feat: enable decimal input on all stocktake screens (store, bar, kitchen)
- EditableCell: value/onChanged double?, keyboard decimal, allow regex formatter
- StockTakeItem: physicalCount double?, variance returns double
- StockTakeProvider: updatePhysicalCount(double?), draft/submit parse as double
- StockTable: onPhysicalCountChanged callback double?
- StockTakePage: physicalCount/totalVariance accumulators as double
- SummaryCard: physicalCount/totalVariance double, smart display formatting

+++ famous_gates_app/lib/features/branch_storekeeper/stock_take/models/stock_take_item.dart
@@ -6,7 +6,7 @@ class StockTakeItem {
   final int openingStock;
   final int sales;
   final int sdds;
-  final int? physicalCount;
+  final double? physicalCount;
   final String? reason;
   final String? explanation;
   final String? actionTaken;
@@ -33,7 +33,7 @@ class StockTakeItem {
 
   int get closingStock => total - sales;
 
-  int get variance => (physicalCount ?? 0) - closingStock;
+  double get variance => (physicalCount ?? 0) - closingStock;
 
   static const Object _unset = Object();
 
@@ -61,7 +61,7 @@ class StockTakeItem {
       sdds: sdds ?? this.sdds,
       physicalCount: identical(physicalCount, _unset)
           ? this.physicalCount
-          : physicalCount as int?,
+          : physicalCount as double?,
       reason: identical(reason, _unset) ? this.reason : reason as String?,
       explanation: identical(explanation, _unset)
           ? this.explanation
